Изменение тестов на отчеты
Изменен вызов функций
